+++ src/RepoAuditor/Plugins/CommunityStandards/Requirements/IssueTemplates.py
@@ -6,7 +6,7 @@
 # -------------------------------------------------------------------------------
 """Contains the IssueTemplates requirement."""
 
-from RepoAuditor.Plugins.GitHubCommunityStandards.Impl.ExistsRequirementImpl import ExistsRequirementImpl
+from RepoAuditor.Plugins.CommunityStandards.Impl.ExistsRequirementImpl import ExistsRequirementImpl
 
 
 class IssueTemplates(ExistsRequirementImpl):
