+++ src/RepoAuditor/Plugins/CommunityStandards/Requirements/SecurityPolicy.py
@@ -8,7 +8,7 @@
 
 import textwrap
 
-from RepoAuditor.Plugins.GitHubCommunityStandards.Impl.ExistsRequirementImpl import ExistsRequirementImpl
+from RepoAuditor.Plugins.CommunityStandards.Impl.ExistsRequirementImpl import ExistsRequirementImpl
 
 
 class SecurityPolicy(ExistsRequirementImpl):
